broker_console/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
paho-mqtt==1.6.1
